refactor: move Dial9Handle and SharedState into dial9-core
