+++ dial9-core/src/lib.rs
@@ -6,6 +6,12 @@ pub mod clock;
 pub mod collector;
 /// Wire-format events emitted by the bus itself.
 pub mod format;
+/// Cloneable handle for recording events and controlling telemetry.
+pub mod handle;
+/// Operational metrics for the core flush path.
+pub mod metrics;
+/// Runtime-agnostic recording state shared across threads.
+pub mod shared_state;
 /// Cfg-gated concurrency primitives (std / shuttle).
 ///
 /// `pub` so sibling crates share one shuttle shim, but not part of the public
