+++ dial9-tokio-telemetry/src/metrics.rs
@@ -66,26 +66,7 @@ pub(crate) struct WorkerCycleMetrics {
     pub segments_dispatched: u64,
 }
 
-/// Per-cycle counters produced by the intrusive thread-local buffer
-/// drain. Also used as a `#[metrics(subfield)]` so callers can flatten
-/// these fields into their top-level metrics without duplication.
-#[metrics(subfield)]
-#[derive(Debug, Default)]
-pub(crate) struct TlDrainStats {
-    /// Buffers that we locked cross-thread and had pending events.
-    pub buffers_flushed: u64,
-    /// Buffers that we locked cross-thread (superset of `buffers_flushed`;
-    /// the difference is buffers that were already empty when locked).
-    pub buffers_locked: u64,
-    /// Handles skipped because the owning thread self-flushed during the
-    /// epoch grace period. High ratio means busy workers are self-flushing
-    /// efficiently and the intrusive path is staying out of their way.
-    pub buffers_skipped_busy: u64,
-    /// Total events drained from idle/silent buffers this cycle.
-    pub events_flushed: u64,
-    /// Dead `Weak` handles pruned this cycle (threads that have exited).
-    pub dead_pruned: u64,
-}
+pub(crate) use dial9_core::metrics::TlDrainStats;
 
 /// Metrics emitted every time the flush thread runs the intrusive
 /// thread-local buffer drain (~every 30s, plus on shutdown).
