--- dial9-tokio-telemetry/src/telemetry/recorder/shared_state.rs
@@ -1,837 +1,0 @@
-use crate::metrics::TlDrainStats;
-use crate::primitives::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use crate::primitives::sync::{Arc, Mutex};
-use crate::telemetry::buffer;
-use crate::telemetry::buffer::TlBufferHandle;
-use crate::telemetry::collector::CentralCollector;
-use std::cell::Cell;
-use std::time::Duration;
-
-crate::primitives::thread_local! {
-    /// schedstat wait_time_ns captured at park time, used to compute delta on unpark.
-    pub(super) static PARKED_SCHED_WAIT: Cell<u64> = const { Cell::new(0) };
-}
-
-/// Runtime-agnostic core recording state.
-///
-/// No tokio imports. All runtime-specific logic lives in `RuntimeContext`.
-pub(crate) struct SharedState {
-    pub(crate) enabled: AtomicBool,
-    /// Set when `TaskDumpConfig` is provided at build time. When `true`,
-    /// wrapping futures capture async backtraces at yield points.
-    pub(crate) task_dumps_enabled: AtomicBool,
-    /// Snapshot of `TaskDumpConfig::idle_threshold` in nanos. Copied onto
-    /// each `TaskDumped` instance at construction time so the hot poll path
-    /// does not need an atomic load.
-    pub(crate) task_dump_idle_threshold_ns: AtomicU64,
-    /// Fixed RNG seed for deterministic task dump sampling. Set once at
-    /// construction before the `Arc` is shared; read-only thereafter.
-    #[cfg_attr(not(feature = "taskdump"), allow(dead_code))]
-    pub(crate) task_dump_rng_seed: Option<u64>,
-    pub(crate) collector: Arc<CentralCollector>,
-    /// Absolute `CLOCK_MONOTONIC` nanosecond timestamp captured at trace start.
-    pub(crate) start_time_ns: u64,
-    /// Global worker ID counter. Each runtime reserves a contiguous block
-    /// via `fetch_add(num_workers)` so worker IDs don't collide.
-    pub(crate) next_worker_id: AtomicU64,
-    /// Epoch counter bumped by the flush thread every ~30s. Thread-local
-    /// buffers stamp this value on each self-flush so the flush thread can
-    /// skip busy workers when draining.
-    pub(crate) drain_epoch: AtomicU64,
-    /// Weak handles to all registered thread-local buffers. The flush thread
-    /// uses these to intrusively drain idle/silent buffers.
-    tl_buffers: Mutex<Vec<TlBufferHandle>>,
-    /// Data sources (CPU profiler, sched profiler, etc.) that the flush thread drains.
-    pub(crate) sources: Mutex<Vec<Box<dyn super::source::Source>>>,
-}
-
-impl SharedState {
-    pub(crate) fn new(start_time_ns: u64, task_dump_rng_seed: Option<u64>) -> Self {
-        Self {
-            enabled: AtomicBool::new(false),
-            task_dumps_enabled: AtomicBool::new(false),
-            task_dump_idle_threshold_ns: AtomicU64::new(0),
-            task_dump_rng_seed,
-            collector: Arc::new(CentralCollector::new()),
-            start_time_ns,
-            next_worker_id: AtomicU64::new(0),
-            drain_epoch: AtomicU64::new(0),
-            tl_buffers: Mutex::new(Vec::new()),
-            sources: Mutex::new(Vec::new()),
-        }
-    }
-
-    /// Register a data source to be drained by the flush thread each cycle.
-    pub(crate) fn push_source(&self, source: Box<dyn super::source::Source>) {
-        self.sources.lock().unwrap().push(source);
-    }
-
-    /// Check whether recording is currently enabled.
-    ///
-    /// Prefer [`if_enabled`](Self::if_enabled) for event-recording paths — it
-    /// provides an [`EventBuffer`] that makes it structurally impossible to
-    /// record without checking first. Use `is_enabled()` only for
-    /// control-flow decisions that don't directly record events (e.g.
-    /// deciding whether to wrap a waker in wake-tracking polls).
-    pub(crate) fn is_enabled(&self) -> bool {
-        self.enabled.load(Ordering::Relaxed)
-    }
-
-    /// Run `f` only when recording is enabled, passing an [`EventBuffer`]
-    /// that provides `record_event` / `record_encodable_event`. Returns
-    /// `None` when disabled (no work is done).
-    pub(crate) fn if_enabled<R>(&self, f: impl FnOnce(&EventBuffer<'_>) -> R) -> Option<R> {
-        if !self.enabled.load(Ordering::Relaxed) {
-            return None;
-        }
-        Some(f(&EventBuffer(self)))
-    }
-
-    /// Test-only shortcut to record an event directly. Production code records
-    /// through [`EventBuffer`] via [`if_enabled`](Self::if_enabled).
-    #[cfg(test)]
-    fn record_encodable_event(&self, event: &dyn buffer::Encodable) {
-        if let Some(handle) =
-            buffer::record_encodable_event(event, &self.collector, &self.drain_epoch)
-        {
-            self.tl_buffers.lock().unwrap().push(handle);
-        }
-    }
-
-    /// Bump the drain epoch and flush all idle/silent thread-local buffers.
-    ///
-    /// Buffers whose `FlushEpoch` matches the current epoch are skipped
-    /// (the owning thread flushed recently, so locking would just add
-    /// contention). Dead `Weak` handles are pruned.
-    ///
-    /// [`bump_drain_epoch`] is called one flush-loop tick
-    /// before calling this method. That gives busy worker threads a ~5 ms
-    /// grace period to self-flush on their next `record_event`, so the
-    /// intrusive drain only needs to lock truly idle/silent buffers.
-    ///
-    /// Returns per-cycle counters so the flush thread can emit metrics.
-    pub(crate) fn drain_all_tl_buffers(&self) -> TlDrainStats {
-        let mut stats = TlDrainStats::default();
-        let epoch = self.drain_epoch.load(Ordering::Relaxed);
-
-        let handles: Vec<TlBufferHandle> = {
-            let guard = self.tl_buffers.lock().unwrap();
-            guard
-                .iter()
-                .map(|h| TlBufferHandle {
-                    buffer: h.buffer.clone(),
-                    flush_epoch: h.flush_epoch.clone(),
-                })
-                .collect()
-        };
-
-        for handle in &handles {
-            // Skip buffers that self-flushed during the current epoch.
-            if handle.flush_epoch.load() >= epoch {
-                stats.buffers_skipped_busy += 1;
-                continue;
-            }
-            if let Some(arc) = handle.buffer.upgrade() {
-                let mut buf = match arc.lock() {
-                    Ok(guard) => guard,
-                    // Buffer is poisoned (encoder panic); skip rather than
-                    // flushing potentially corrupt data.
-                    Err(_) => {
-                        crate::rate_limit::rate_limited!(Duration::from_secs(60), {
-                            tracing::error!(
-                                "dial9: thread-local buffer mutex poisoned in drain_all_tl_buffers; skipping flush"
-                            );
-                        });
-                        continue;
-                    }
-                };
-                stats.buffers_locked += 1;
-                if buf.has_pending_events() {
-                    let batch = buf.flush();
-                    stats.events_flushed += batch.event_count();
-                    stats.buffers_flushed += 1;
-                    self.collector.accept_flush(batch);
-                }
-                // Stamp so we skip this buffer next cycle if it stays idle.
-                handle.flush_epoch.store(epoch);
-            }
-        }
-
-        // Prune dead handles (Weak refs to threads that have exited).
-        let mut guard = self.tl_buffers.lock().unwrap();
-        let before = guard.len();
-        guard.retain(|h| h.buffer.strong_count() > 0);
-        stats.dead_pruned = (before - guard.len()) as u64;
-
-        stats
-    }
-
-    /// Advance the global drain epoch so that busy worker threads
-    /// self-flush on their next `record_event` call. Call this one
-    /// flush-loop tick (~5 ms) before [`drain_all_tl_buffers`] to give
-    /// workers a grace period, minimising contention on the intrusive
-    /// drain path.
-    pub(crate) fn bump_drain_epoch(&self) {
-        self.drain_epoch.fetch_add(1, Ordering::Relaxed);
-    }
-
-    /// Drain data sources and write their events into the collector.
-    pub(crate) fn flush_sources(&self) {
-        use super::source::FlushContext;
-
-        let ctx = FlushContext {
-            collector: &self.collector,
-            drain_epoch: &self.drain_epoch,
-        };
-
-        let mut sources = self.sources.lock().unwrap();
-        for source in sources.iter_mut() {
-            source.flush(&ctx);
-        }
-    }
-}
-
-/// Handle provided by [`SharedState::if_enabled`] that proves recording is
-/// active. All event-recording calls should go through this type so that
-/// callers cannot accidentally emit events without an enabled check.
-pub(crate) struct EventBuffer<'a>(&'a SharedState);
-
-impl EventBuffer<'_> {
-    pub(crate) fn record_encodable_event(&self, event: &dyn buffer::Encodable) {
-        if let Some(handle) =
-            buffer::record_encodable_event(event, &self.0.collector, &self.0.drain_epoch)
-        {
-            self.0.tl_buffers.lock().unwrap().push(handle);
-        }
-    }
-
-    pub(crate) fn with_encoder(&self, f: impl FnOnce(&mut buffer::ThreadLocalEncoder<'_>)) {
-        if let Some(handle) = buffer::with_encoder(f, &self.0.collector, &self.0.drain_epoch) {
-            self.0.tl_buffers.lock().unwrap().push(handle);
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::telemetry::format::{PollEndEvent, WorkerId};
-
-    fn poll_end_event() -> PollEndEvent {
-        PollEndEvent {
-            timestamp_ns: 1000,
-            worker_id: WorkerId::from(0usize),
-        }
-    }
-
-    /// Helper: create a SharedState with recording enabled.
-    fn enabled_shared_state() -> SharedState {
-        let ss = SharedState::new(0, None);
-        ss.enabled.store(true, Ordering::Relaxed);
-        ss
-    }
-
-    #[test]
-    fn record_event_registers_tl_buffer_handle() {
-        let ss = enabled_shared_state();
-        // First event on this thread should register a handle.
-        ss.record_encodable_event(&poll_end_event());
-        let handles = ss.tl_buffers.lock().unwrap();
-        assert_eq!(handles.len(), 1);
-        assert!(handles[0].buffer.upgrade().is_some());
-    }
-
-    #[test]
-    fn second_record_event_does_not_re_register() {
-        let ss = enabled_shared_state();
-        ss.record_encodable_event(&poll_end_event());
-        ss.record_encodable_event(&poll_end_event());
-        let handles = ss.tl_buffers.lock().unwrap();
-        assert_eq!(handles.len(), 1);
-    }
-
-    #[test]
-    fn drain_all_tl_buffers_flushes_idle_buffer() {
-        let ss = enabled_shared_state();
-        // Write an event (won't self-flush — buffer is 1MB).
-        ss.record_encodable_event(&poll_end_event());
-        // Nothing in the collector yet (buffer not full).
-        assert!(ss.collector.next().is_none());
-        // Bump epoch so the idle buffer (epoch 0) is stale, then drain.
-        ss.bump_drain_epoch();
-        ss.drain_all_tl_buffers();
-        let batch = ss.collector.next().expect("expected a batch after drain");
-        assert!(batch.event_count() > 0);
-    }
-
-    #[test]
-    fn drain_all_tl_buffers_from_another_thread() {
-        let ss = Arc::new(enabled_shared_state());
-        let ss2 = ss.clone();
-        // Write events from a spawned thread.
-        let handle = std::thread::spawn(move || {
-            ss2.record_encodable_event(&poll_end_event());
-            ss2.record_encodable_event(&poll_end_event());
-        });
-        handle.join().unwrap();
-        // Bump epoch so the buffer is stale, then drain from the main thread.
-        ss.bump_drain_epoch();
-        ss.drain_all_tl_buffers();
-        let batch = ss.collector.next().expect("expected a batch after drain");
-        assert_eq!(batch.event_count(), 2);
-    }
-
-    #[test]
-    fn drain_skips_busy_buffer() {
-        let ss = enabled_shared_state();
-        ss.record_encodable_event(&poll_end_event());
-        // Bump epoch to 1 (simulates the tick before the drain).
-        ss.bump_drain_epoch();
-        // Simulate a self-flush by stamping the current epoch.
-        {
-            let handles = ss.tl_buffers.lock().unwrap();
-            handles[0].flush_epoch.store(1);
-        }
-        ss.drain_all_tl_buffers();
-        // Buffer should NOT have been flushed — collector is empty.
-        assert!(ss.collector.next().is_none());
-    }
-
-    #[test]
-    fn drain_prunes_dead_handles() {
-        let ss = Arc::new(enabled_shared_state());
-        let ss2 = ss.clone();
-        let handle = std::thread::spawn(move || {
-            ss2.record_encodable_event(&poll_end_event());
-        });
-        handle.join().unwrap();
-        // Thread exited — its Arc<Mutex<TLB>> was dropped, Weak is dead.
-        // But the TLB's Drop impl flushed remaining events, so the handle
-        // is dead. Drain should prune it.
-        ss.drain_all_tl_buffers();
-        let handles = ss.tl_buffers.lock().unwrap();
-        assert_eq!(handles.len(), 0, "dead handle should have been pruned");
-    }
-
-    /// Intrusive-drain path with a *live* worker thread. Unlike
-    /// `drain_all_tl_buffers_from_another_thread`, which joins the worker
-    /// before draining (so events reach the collector via the TLB `Drop`
-    /// impl, not via the intrusive path), here the worker is parked on a
-    /// channel while the main thread bumps+drains, proving that
-    /// `drain_all_tl_buffers` upgrades the live `Weak`, locks the mutex
-    /// cross-thread, and flushes the pending event.
-    #[test]
-    fn drain_flushes_live_worker_buffer() {
-        let ss = Arc::new(enabled_shared_state());
-        let ss2 = ss.clone();
-        let (release_tx, release_rx) = std::sync::mpsc::channel::<()>();
-        let (ready_tx, ready_rx) = std::sync::mpsc::channel::<()>();
-
-        let worker = std::thread::spawn(move || {
-            // drain_epoch is 0, so no self-flush happens — the event
-            // stays in the buffer.
-            ss2.record_encodable_event(&poll_end_event());
-            ready_tx.send(()).unwrap();
-            // Park until main thread has drained. The TLB `Drop` impl must
-            // not run before the intrusive drain, otherwise we're not
-            // testing the intrusive path.
-            release_rx.recv().unwrap();
-        });
-
-        ready_rx.recv().unwrap();
-        // Worker is parked with one event in its TLB and a live handle.
-        // Nothing in the collector yet — no self-flush was triggered.
-        assert!(ss.collector.next().is_none());
-
-        ss.bump_drain_epoch();
-        ss.drain_all_tl_buffers();
-
-        let batch = ss
-            .collector
-            .next()
-            .expect("intrusive drain should have flushed the live worker's event");
-        assert_eq!(batch.event_count(), 1);
-
-        release_tx.send(()).unwrap();
-        worker.join().unwrap();
-    }
-
-    // Concurrent-stress proptest: the core invariant of the TL buffer
-    // drain feature is that no events are lost and none are duplicated,
-    // regardless of how `record_encodable_event`, `bump_drain_epoch`, and
-    // `drain_all_tl_buffers` interleave across threads. Spawn N writer
-    // threads, each recording M events, while a drainer thread
-    // concurrently bumps+drains. After joining, a final bump+drain should
-    // leave exactly N*M events in the collector.
-    proptest::proptest! {
-        #![proptest_config(proptest::prelude::ProptestConfig::with_cases(32))]
-
-        #[test]
-        fn concurrent_record_and_drain_preserves_event_count(
-            num_threads in 1usize..=6,
-            events_per_thread in 1u64..=200,
-            drain_ticks in 0usize..=10,
-        ) {
-            let ss = Arc::new(enabled_shared_state());
-            let start = Arc::new(std::sync::Barrier::new(num_threads + 1));
-            let stop_drainer = Arc::new(AtomicBool::new(false));
-
-            let writers: Vec<_> = (0..num_threads)
-                .map(|_| {
-                    let ss = ss.clone();
-                    let start = start.clone();
-                    std::thread::spawn(move || {
-                        start.wait();
-                        for _ in 0..events_per_thread {
-                            ss.record_encodable_event(&poll_end_event());
-                        }
-                    })
-                })
-                .collect();
-
-            let drainer = {
-                let ss = ss.clone();
-                let stop = stop_drainer.clone();
-                std::thread::spawn(move || {
-                    let mut ticks = 0;
-                    while ticks < drain_ticks && !stop.load(Ordering::Relaxed) {
-                        ss.bump_drain_epoch();
-                        // Short grace period so any in-flight writer has a
-                        // chance to self-flush before the intrusive drain.
-                        std::thread::sleep(std::time::Duration::from_micros(50));
-                        ss.drain_all_tl_buffers();
-                        ticks += 1;
-                    }
-                })
-            };
-
-            start.wait();
-            for w in writers {
-                w.join().unwrap();
-            }
-            stop_drainer.store(true, Ordering::Relaxed);
-            drainer.join().unwrap();
-
-            // Writer threads have exited, so their TLB `Drop` impls have
-            // flushed any remaining events. Do one final bump+drain to
-            // prune dead handles (no-op for event capture at this point).
-            ss.bump_drain_epoch();
-            ss.drain_all_tl_buffers();
-
-            let mut total: u64 = 0;
-            while let Some(batch) = ss.collector.next() {
-                total += batch.event_count();
-            }
-            // Sanity: the collector never evicted a batch under these
-            // workloads. If it did, the invariant check below would be
-            // meaningless.
-            proptest::prop_assert_eq!(ss.collector.take_dropped_batches(), 0);
-            proptest::prop_assert_eq!(
-                total,
-                num_threads as u64 * events_per_thread,
-                "every recorded event must reach the collector exactly once"
-            );
-        }
-    }
-}
-
-#[cfg(all(test, shuttle))]
-mod shuttle_tests {
-    use crate::primitives::fs;
-    use crate::primitives::sync::atomic::{AtomicU64, Ordering};
-    use crate::primitives::sync::{Arc, Mutex};
-    use crate::telemetry::recorder::TelemetryCore;
-    use crate::telemetry::writer::{DiskWriter, InMemoryWriter};
-    use dial9_trace_format::TraceEvent;
-    use shuttle::rand::Rng;
-    use std::collections::HashMap;
-
-    // ── Event definition ────────────────────────────────────────────────
-
-    /// Custom event for round-trip validation. Each event carries a
-    /// per-thread monotonic `seq`, a `thread_id`, and a `timestamp_ns`
-    /// that is mostly monotonic with occasional backward jumps.
-    #[derive(TraceEvent, Clone, Debug, serde::Deserialize)]
-    struct ValidationEvent {
-        #[traceevent(timestamp)]
-        timestamp_ns: u64,
-        thread_id: u64,
-        seq: u64,
-        id: u64,
-    }
-
-    /// Generate a timestamp that is mostly monotonic with occasional backward
-    /// jumps, driven by shuttle's deterministic RNG.
-    fn next_timestamp(prev: &mut u64) -> u64 {
-        let mut rng = shuttle::rand::thread_rng();
-        if rng.gen_range(0u32..5) == 0 {
-            *prev = prev.saturating_sub(rng.gen_range(1u64..=100));
-        } else {
-            *prev += rng.gen_range(1u64..=1000);
-        }
-        *prev
-    }
-
-    // ── Decoding ────────────────────────────────────────────────────────
-
-    fn decode_validation_events(data: &[u8]) -> Vec<ValidationEvent> {
-        use dial9_trace_format::decoder::Decoder;
-        let Some(mut dec) = Decoder::new(data) else {
-            assert!(data.is_empty(), "failed to non-empty segment!");
-            return vec![];
-        };
-        let mut out = Vec::new();
-        dec.for_each_event(|ev| {
-            if ev.name == "ValidationEvent" {
-                if let Ok(decoded) = ev.deserialize::<ValidationEvent>() {
-                    out.push(decoded);
-                }
-            }
-        })
-        .expect("decode failed");
-        out
-    }
-
-    // ── Invariants ──────────────────────────────────────────────────────
-
-    /// All emitted event IDs appear exactly once in the decoded output.
-    fn check_all_events_present(expected: &[ValidationEvent], decoded: &[ValidationEvent]) {
-        let mut exp_ids: Vec<u64> = expected.iter().map(|e| e.id).collect();
-        let mut dec_ids: Vec<u64> = decoded.iter().map(|e| e.id).collect();
-        exp_ids.sort();
-        dec_ids.sort();
-        assert_eq!(
-            exp_ids,
-            dec_ids,
-            "event ids mismatch: expected {} events, got {}",
-            exp_ids.len(),
-            dec_ids.len()
-        );
-    }
-
-    /// Every event's timestamp round-trips exactly.
-    fn check_timestamps_roundtrip(expected: &[ValidationEvent], decoded: &[ValidationEvent]) {
-        let exp_by_id: HashMap<u64, u64> =
-            expected.iter().map(|e| (e.id, e.timestamp_ns)).collect();
-        for ev in decoded {
-            let exp_ts = exp_by_id[&ev.id];
-            assert_eq!(
-                exp_ts, ev.timestamp_ns,
-                "timestamp mismatch for event id {}: expected {exp_ts}, got {}",
-                ev.id, ev.timestamp_ns
-            );
-        }
-    }
-
-    // ── Mock Source ────────────────────────────────────────────────────
-
-    /// A Source that accumulates events from worker threads and emits them
-    /// during flush. Exercises the Source flush path under shuttle.
-    struct MockSource {
-        pending: Arc<Mutex<Vec<ValidationEvent>>>,
-    }
-
-    impl MockSource {
-        fn new(pending: Arc<Mutex<Vec<ValidationEvent>>>) -> Self {
-            Self { pending }
-        }
-    }
-
-    impl crate::telemetry::recorder::source::Source for MockSource {
-        fn flush(&mut self, ctx: &crate::telemetry::recorder::source::FlushContext<'_>) {
-            let events: Vec<_> = self.pending.lock().unwrap().drain(..).collect();
-            for ev in &events {
-                crate::telemetry::buffer::record_encodable_event(
-                    ev,
-                    ctx.collector,
-                    ctx.drain_epoch,
-                );
-            }
-        }
-
-        fn name(&self) -> &'static str {
-            "mock"
-        }
-
-        // TODO: exercise on_worker_thread_start/on_thread_stop once shuttle
-        // tests include a Tokio runtime.
-    }
-
-    // ── Test body ───────────────────────────────────────────────────────
-
-    fn test_telemetry_core_pipeline() {
-        let _ts_guard =
-            metrique_timesource::set_time_source(metrique_timesource::TimeSource::custom(
-                metrique_timesource::fakes::StaticTimeSource::at_time(std::time::UNIX_EPOCH),
-            ));
-
-        let num_threads = 3;
-        let next_id = Arc::new(AtomicU64::new(0));
-
-        // Small segments force frequent rotation: the 100 MiB budget is far above the test's data,
-        // so the ring never evicts before we drain it.
-        let writer = InMemoryWriter::builder()
-            .max_total_size(100 * 1024 * 1024)
-            .max_segment_size(256)
-            .build()
-            .unwrap();
-        let fs = writer.fs_handle().expect("in-memory writer exposes its fs");
-
-        // Mock source: worker threads push events here, flush thread drains them.
-        let source_pending: Arc<Mutex<Vec<ValidationEvent>>> = Arc::new(Mutex::new(Vec::new()));
-
-        let guard = TelemetryCore::builder().writer(writer).build().unwrap();
-        guard
-            .shared()
-            .unwrap()
-            .push_source(Box::new(MockSource::new(source_pending.clone())));
-        guard.enable();
-        let handle = guard.handle();
-
-        let expected: Arc<Mutex<Vec<ValidationEvent>>> = Arc::new(Mutex::new(Vec::new()));
-        let writers: Vec<_> = (0..num_threads)
-            .map(|thread_id| {
-                let h = handle.clone();
-                let next_id = next_id.clone();
-                let expected = expected.clone();
-                let source_pending = source_pending.clone();
-                let thread_id = thread_id as u64;
-                crate::primitives::thread::spawn(move || {
-                    let mut rng = shuttle::rand::thread_rng();
-                    let count = rng.gen_range(3u64..=10);
-                    let mut ts = rng.gen_range(1000u64..2000);
-                    for seq in 0..count {
-                        let id = next_id.fetch_add(1, Ordering::Relaxed);
-                        let timestamp_ns = next_timestamp(&mut ts);
-                        let ev = ValidationEvent {
-                            timestamp_ns,
-                            thread_id,
-                            seq,
-                            id,
-                        };
-                        expected.lock().unwrap().push(ev.clone());
-                        // Randomly choose: emit via handle (TL buffer path)
-                        // or via mock source (flush-thread path).
-                        if rng.gen_range(0u32..2) == 0 {
-                            h.record_encodable_event(&ev);
-                        } else {
-                            source_pending.lock().unwrap().push(ev);
-                        }
-                    }
-                })
-            })
-            .collect();
-
-        for w in writers {
-            w.join().unwrap();
-        }
-        drop(guard);
-
-        // Drain the in-memory ring (memory pops one sealed segment per call).
-        let mut all_decoded: Vec<ValidationEvent> = Vec::new();
-        loop {
-            let taken = fs.take_files();
-            if taken.segments.is_empty() {
-                break;
-            }
-            for seg in taken.segments {
-                let (_seg_ref, payload, _accounting) = seg.load().unwrap();
-                all_decoded.extend(decode_validation_events(&payload.into_vec()));
-            }
-        }
-        let expected = expected.lock().unwrap();
-
-        // Run all invariants.
-        check_all_events_present(&expected, &all_decoded);
-        check_timestamps_roundtrip(&expected, &all_decoded);
-    }
-
-    #[test]
-    fn determinism_check() {
-        shuttle::check_uncontrolled_nondeterminism(test_telemetry_core_pipeline, 10000);
-    }
-
-    #[test]
-    fn pct_real_pipeline() {
-        shuttle::check_pct(test_telemetry_core_pipeline, 10000, 3);
-    }
-
-    // ── Error injection ─────────────────────────────────────────────────
-    //
-    // Companion to the round-trip pipeline test. `primitives::fs` is armed
-    // with a fault policy so the writer's real I/O points (write/flush/rename/
-    // remove) fail, exercising the flush loop's error paths.
-
-    use std::sync::Arc as StdArc;
-    use std::sync::atomic::{AtomicU64 as StdAtomicU64, Ordering as StdOrdering};
-
-    // ── Counting subscriber ─────────────────────────────────────────────
-    //
-    // A minimal `tracing::Subscriber` that increments a shared counter
-    // on every WARN or ERROR event. We use `tracing::subscriber::with_default`
-    // to scope it to a single test invocation. We deliberately avoid
-    // depending on `tracing-subscriber` since that crate is gated behind
-    // the `tracing-layer` feature and isn't enabled under `_shuttle`.
-
-    struct CountingSubscriber {
-        warn_or_error_count: StdArc<StdAtomicU64>,
-    }
-
-    impl tracing::Subscriber for CountingSubscriber {
-        fn enabled(&self, metadata: &tracing::Metadata<'_>) -> bool {
-            matches!(
-                *metadata.level(),
-                tracing::Level::WARN | tracing::Level::ERROR
-            )
-        }
-
-        fn new_span(&self, _span: &tracing::span::Attributes<'_>) -> tracing::span::Id {
-            // We never actually use spans; return a fixed non-zero id.
-            tracing::span::Id::from_u64(1)
-        }
-
-        fn record(&self, _span: &tracing::span::Id, _values: &tracing::span::Record<'_>) {}
-
-        fn record_follows_from(&self, _span: &tracing::span::Id, _follows: &tracing::span::Id) {}
-
-        fn event(&self, event: &tracing::Event<'_>) {
-            let level = *event.metadata().level();
-            if level == tracing::Level::WARN || level == tracing::Level::ERROR {
-                self.warn_or_error_count.fetch_add(1, StdOrdering::Relaxed);
-            }
-        }
-
-        fn enter(&self, _span: &tracing::span::Id) {}
-
-        fn exit(&self, _span: &tracing::span::Id) {}
-    }
-
-    // ── Erroring-pipeline test body ─────────────────────────────────────
-
-    /// Drive the pipeline with the fs armed to `fault`, returning the
-    /// number of WARN/ERROR events the flush loop emitted.
-    fn run_erroring_pipeline(fault: fs::FaultPolicy) -> u64 {
-        let _ts_guard =
-            metrique_timesource::set_time_source(metrique_timesource::TimeSource::custom(
-                metrique_timesource::fakes::StaticTimeSource::at_time(std::time::UNIX_EPOCH),
-            ));
-
-        let warn_count = StdArc::new(StdAtomicU64::new(0));
-        let subscriber = CountingSubscriber {
-            warn_or_error_count: warn_count.clone(),
-        };
-
-        tracing::subscriber::with_default(subscriber, || {
-            let num_threads = 3;
-            let next_id = Arc::new(AtomicU64::new(0));
-
-            let dir = tempfile::tempdir().unwrap();
-            let writer = DiskWriter::single_file(dir.path().join("trace.bin")).unwrap();
-            let _fault = fs::set_fault(fault);
-            let guard = TelemetryCore::builder().writer(writer).build().unwrap();
-            guard.enable();
-            let handle = guard.handle();
-
-            let writers: Vec<_> = (0..num_threads)
-                .map(|thread_id| {
-                    let h = handle.clone();
-                    let next_id = next_id.clone();
-                    let thread_id = thread_id as u64;
-                    crate::primitives::thread::spawn(move || {
-                        let mut rng = shuttle::rand::thread_rng();
-                        let count = rng.gen_range(3u64..=10);
-                        let mut ts = rng.gen_range(1000u64..2000);
-                        for seq in 0..count {
-                            let id = next_id.fetch_add(1, Ordering::Relaxed);
-                            let timestamp_ns = next_timestamp(&mut ts);
-                            let ev = ValidationEvent {
-                                timestamp_ns,
-                                thread_id,
-                                seq,
-                                id,
-                            };
-                            // Errors are expected; whether the event is
-                            // recorded or dropped is not asserted here.
-                            h.record_encodable_event(&ev);
-                        }
-                    })
-                })
-                .collect();
-
-            for w in writers {
-                w.join().unwrap();
-            }
-            // Dropping the guard triggers the shutdown/finalize path,
-            // which should also be rate-limited if it logs on error.
-            drop(guard);
-        });
-
-        warn_count.load(StdOrdering::Relaxed)
-    }
-
-    /// The fault is armed on the test thread but read on the flush thread,
-    /// this proves it crosses that boundary.
-    fn fs_fault_visible_across_threads() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("fault_probe");
-        std::fs::write(&path, b"x").unwrap();
-
-        let _fault = fs::set_fault(fs::FaultPolicy::FailAll);
-        let observed_fault =
-            crate::primitives::thread::spawn(move || fs::remove_file(&path).is_err())
-                .join()
-                .unwrap();
-
-        assert!(
-            observed_fault,
-            "fault armed on the test thread was not observed on a spawned thread"
-        );
-    }
-
-    #[test]
-    fn determinism_check_fs_fault_visible() {
-        shuttle::check_uncontrolled_nondeterminism(fs_fault_visible_across_threads, 100);
-    }
-
-    fn test_telemetry_core_erroring_pipeline() {
-        let total = run_erroring_pipeline(fs::FaultPolicy::FailAll);
-        assert!(
-            total <= 10,
-            "rate limiting failed under persistent writer errors: \
-             observed {total} WARN/ERROR events, expected <= 10. \
-             A `rate_limited!` wrapper has likely been removed from a tight loop."
-        );
-    }
-
-    #[test]
-    fn determinism_check_erroring() {
-        shuttle::check_uncontrolled_nondeterminism(test_telemetry_core_erroring_pipeline, 10000);
-    }
-
-    #[test]
-    fn pct_erroring_pipeline() {
-        shuttle::check_pct(test_telemetry_core_erroring_pipeline, 10000, 3);
-    }
-
-    fn test_telemetry_core_probabilistic_fs_faults() {
-        let total = run_erroring_pipeline(fs::FaultPolicy::FailProb(0.5));
-        assert!(
-            total <= 10,
-            "rate limiting failed under probabilistic fs faults: observed {total} \
-             WARN/ERROR events, expected <= 10."
-        );
-    }
-
-    #[test]
-    fn determinism_check_probabilistic_fs_faults() {
-        shuttle::check_uncontrolled_nondeterminism(
-            test_telemetry_core_probabilistic_fs_faults,
-            10000,
-        );
-    }
-
-    #[test]
-    fn pct_probabilistic_fs_faults() {
-        shuttle::check_pct(test_telemetry_core_probabilistic_fs_faults, 10000, 3);
-    }
-}
